Don't use "@overload" to document methods without parameters.
Documentation generated by @overload is misleading in this case as it refers
to the "different argument [the function] accepts", but there are none, the
methods can be overloaded only by const modifier as well.

Use \copydetails instead to reuse the original documentation, this is less
confusing and arguably more useful as well.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@78078 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775