Fix best size for non-left aligned wxStaticText in wxGTK2
Since the hack for making the alignment work with GTK+ 2 done in
aa6b2207420ccece6d6b71452d86c2b546dbcfc7, the best size was not computed
correctly for the right aligned or centered labels and the minimum possible
size was returned instead.

Fix this by temporarily disabling ellipsization during the best size
computation to ensure that the same best size is computed for left and right
aligned labels.

See #12539.